// RRDeclarat Proyecto1/src/types/interfaces.ts
// PokemonListResponse is the response of the API call https://pokeapi.co/api/v2/pokemon/?limit=2000
//definimos un tipo para el estado de carga de la pagina
export type estadoCarga = "inicio" | "cargando" | "completo" | string; //el error lo metemos en el string, porque acompañará el mensaje de error



export interface PokemonBasic {
  name: string;
  url: string;
}

export interface PokemonListResponse {
  count: number;
  next: string | null;
  previous: string | null;
  results: PokemonBasic[];
}

// Pokemon is the response of the API call https://pokeapi.co/api/v2/pokemon/{name}

export interface Pokemon {
  abilities: AbilitySlot[];
  base_experience: number;
  cries: Cries;
  forms: NamedAPIResource[];
  game_indices: GameIndex[];
  height: number;
  held_items: HeldItem[];
  id: number;
  is_default: boolean;
  location_area_encounters: string;
  moves: Move[];
  name: string;
  order: number;
  past_abilities: unknown[]; // No hay datos en Ditto, pero podrían existir en otros Pokémon
  past_types: unknown[];
  species: NamedAPIResource;
  sprites: Sprites;
  stats: Stat[];
  types: TypeSlot[];
  weight: number;
}

interface AbilitySlot {
  ability: NamedAPIResource;
  is_hidden: boolean;
  slot: number;
}

interface Cries {
  latest: string;
  legacy: string;
}

interface NamedAPIResource {
  name: string;
  url: string;
}

interface GameIndex {
  game_index: number;
  version: NamedAPIResource;
}

interface HeldItem {
  item: NamedAPIResource;
  version_details: VersionDetail[];
}

interface VersionDetail {
  rarity: number;
  version: NamedAPIResource;
}

interface Move {
  move: NamedAPIResource;
  version_group_details: VersionGroupDetail[];
}

interface VersionGroupDetail {
  level_learned_at: number;
  move_learn_method: NamedAPIResource;
  version_group: NamedAPIResource;
}

interface Sprites {
  back_default: string | null;
  back_female: string | null;
  back_shiny: string | null;
  back_shiny_female: string | null;
  front_default: string | null;
  front_female: string | null;
  front_shiny: string | null;
  front_shiny_female: string | null;
  other?: {
    dream_world?: {
      front_default: string | null;
      front_female: string | null;
    };
    home?: {
      front_default: string | null;
      front_female: string | null;
      front_shiny: string | null;
      front_shiny_female: string | null;
    };
    official_artwork?: {
      front_default: string | null;
      front_shiny: string | null;
    };
    showdown?: {
      back_default: string | null;
      back_female: string | null;
      back_shiny: string | null;
      back_shiny_female: string | null;
      front_default: string | null;
      front_female: string | null;
      front_shiny: string | null;
      front_shiny_female: string | null;
    };
  };
}

interface Stat {
  base_stat: number;
  effort: number;
  stat: NamedAPIResource;
}

interface TypeSlot {
  slot: number;
  type: NamedAPIResource;
}